requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.3
django-cors-headers==2.4.0
djangorestframework==3.9.0
gspread==3.1.0
gunicorn==19.9.0
httplib2==0.12.0
idna==2.8
oauth2client==4.1.3
Pillow==5.3.0
psycopg2-binary==2.7.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pytz==2018.7
requests==2.21.0
rsa==4.0
six==1.12.0
soupsieve==1.7.3
urllib3==1.24.1
